PPO being added

+++ datasetgenerator.py
@@ -15,11 +15,11 @@ class GridDataset(Dataset):
         return self.data[idx]
 
     def _generate_grid(self):
-        grid = np.ones((self.n, self.n))  # Initialize a 5x5 grid with ones for RGB white cells
+        grid = np.ones((self.n, self.n), dtype=np.float32)  # Initialize a 5x5 grid with ones for RGB white cells
 
         for col in range(self.n-2):
             random_row = np.random.randint(0, self.n)
-            grid[random_row, col + 2] = 0  # Set the chosen cell to blue (assuming [0, 0, 1] is blue in your RGB representation)
+            grid[random_row, col + 2] = 0.0  # Set the chosen cell to blue (assuming [0, 0, 1] is blue in your RGB representation)
         return grid
